requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
appnope==0.1.4; platform_system == "Darwin"
asttokens==3.0.0
async-timeout==5.0.1; python_version < "3.11"
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1; implementation_name == "pypy"
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6; platform_system == "Windows" or sys_platform == "win32"
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
depth-anything-v2 @ git+ssh://git@github.com/debOliveira/depth-anything-V2.git@7885bbc0647bc64d55ff5803561ea2c7dea1af72
diffusers==0.11.1
diffusion-policy @ git+ssh://git@github.com/debOliveira/diffusion_policy.git@db1434cc256b53deb0ad7228c129c0ce7c733822
efficientnet-pytorch==0.7.1
einops==0.8.1
exceptiongroup==1.3.0; python_version < "3.11"
executing==2.2.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec[http]==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
huggingface-hub==0.25.2
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lightning-utilities==0.15.2
lmdb==1.7.3
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-cupti-cu12==12.6.80; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-nvrtc-cu12==12.6.77; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cuda-runtime-cu12==12.6.77; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cudnn-cu12==9.5.1.17; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufft-cu12==11.3.0.4; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cufile-cu12==1.11.1.6; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-curand-cu12==10.3.7.77; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusolver-cu12==11.7.1.2; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparse-cu12==12.5.4.2; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-cusparselt-cu12==0.6.3; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nccl-cu12==2.26.2; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvjitlink-cu12==12.6.85; platform_system == "Linux" and platform_machine == "x86_64"
nvidia-nvtx-cu12==12.6.77; platform_system == "Linux" and platform_machine == "x86_64"
packaging==25.0
pandas==2.3.1
parso==0.8.4
pastel==0.2.1
pexpect==4.9.0; sys_platform != "win32" and sys_platform != "emscripten"
pillow==11.3.0
platformdirs==4.3.8
poethepoet==0.10.0
pot==0.9.5
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1; python_version > "3.9" or sys_platform != "linux"
psutil==7.0.0
ptyprocess==0.7.0; sys_platform != "win32" and sys_platform != "emscripten"
pure-eval==0.2.3
pycparser==2.22; implementation_name == "pypy"
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
pywin32==311; sys_platform == "win32" and platform_python_implementation != "PyPy"
pyyaml==6.0.2
pyzmq==27.0.1
regex==2025.7.34
requests==2.32.4
scikit-learn==1.7.1
scipy==1.15.3
sentry-sdk==2.34.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.7.1
torchcde==0.2.5
torchcfm==1.0.7
torchdiffeq==0.2.5
torchdyn==1.0.6
torchmetrics==1.8.1
torchsde==0.2.6
torchvision==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
triton==3.3.1; platform_system == "Linux" and platform_machine == "x86_64"
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.1
warmup-scheduler==0.3
wcwidth==0.2.13
widgetsnbextension==4.0.14
yarl==1.20.1
zipp==3.23.0
cv-bridge==3.2.1